Test for Location sources.
When fingerprinting a file, it is necessary to mark what document / source a
fingerprint came from. This way, if a user wants to find matches between two
collections of documents `a` and `b`, they could do:

  fingerprinter = Winnow::Fingerprinter.new(...)

  fingerprints_a = a.map do |doc|
    fingerprinter.fingerprints(doc.to_s, doc.name)
  end.reduce(:union)

  fingerprints_b = b.map do |doc|
    fingerprinter.fingerprints(doc.to_s, doc.name)
  end.reduce(:union)

  Winnow::Matcher.find_matches(a, b).each do |match|
    puts "Matches from doc a are found in files called:"
    puts match.matches_from_a.map { |fp| fp.location.source }.join(", ")

    puts "Matches from doc b are found in files called:"
    puts match.matches_from_b.map { |fp| fp.location.source }.join(", ")
  end

This seems complicated, but it's a necessary evil to handle the rather
complicated general-case situations involved with plagiarism detection.